log fixed

+++ dual_core_completo/dual_core/cpu2_control_unit_test/car_management.h
@@ -61,7 +61,7 @@ extern int throttleReq;
 
 extern float accelerations[3];
 extern float omegas[3];
-extern Uint32 suspensions[4];
+extern float suspensions[4];
 extern float temperatures[8];
 
 extern float bms_lv_cell[8];
